re-use same rollbar provider code in generic issuance client and zupoll
