+++ apps/generic-issuance-client/package.json
@@ -18,6 +18,7 @@
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@monaco-editor/react": "^4.6.0",
+    "@pcd/client-shared": "0.0.1",
     "@pcd/passport-interface": "0.11.0",
     "@pcd/util": "0.5.0",
     "@rollbar/react": "^0.11.2",
