+++ apps/generic-issuance-client/tsconfig.json
@@ -3,7 +3,10 @@
     "resolveJsonModule": true,
     "downlevelIteration": true,
     "jsx": "react-jsx",
-    "lib": ["ES2020", "DOM"],
+    "lib": [
+      "ES2020",
+      "DOM"
+    ],
     "esModuleInterop": true,
     "sourceMap": true,
     // To allow for mocha and jest to work together:
@@ -11,9 +14,17 @@
     "skipLibCheck": true,
     "strictNullChecks": true
   },
-  "include": ["**/*.ts", "**/*.tsx"],
-  "exclude": ["node_modules"],
+  "include": [
+    "**/*.ts",
+    "**/*.tsx"
+  ],
+  "exclude": [
+    "node_modules"
+  ],
   "references": [
+    {
+      "path": "../../packages/lib/client-shared"
+    },
     {
       "path": "../../packages/lib/passport-interface"
     },
