+++ apps/passport-server/tsconfig.json
@@ -60,10 +60,10 @@
       "path": "../../packages/lib/pod"
     },
     {
-      "path": "../../packages/pcd/pod-ticket-pcd"
+      "path": "../../packages/pcd/pod-pcd"
     },
     {
-      "path": "../../packages/pcd/pod-pcd"
+      "path": "../../packages/pcd/pod-ticket-pcd"
     },
     {
       "path": "../../packages/pcd/rln-pcd"
