+++ apps/zupoll-client/tsconfig.json
@@ -1,7 +1,14 @@
 {
   "extends": "@pcd/tsconfig/nextjs.json",
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-  "exclude": ["node_modules"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts"
+  ],
+  "exclude": [
+    "node_modules"
+  ],
   "compilerOptions": {
     "jsx": "preserve",
     "moduleResolution": "Bundler",
@@ -15,10 +22,15 @@
     "strictNullChecks": true,
     "strict": true,
     "paths": {
-      "@/*": ["./src/@/*"]
+      "@/*": [
+        "./src/@/*"
+      ]
     }
   },
   "references": [
+    {
+      "path": "../../packages/lib/client-shared"
+    },
     {
       "path": "../../packages/lib/passport-interface"
     },
