+++ tsconfig.cjs.json
@@ -4,6 +4,9 @@
   // References are automatically maintained by `yarn fix-references`
   "references": [
     {
+      "path": "packages/lib/client-shared/tsconfig.cjs.json"
+    },
+    {
       "path": "packages/lib/emitter/tsconfig.cjs.json"
     },
     {
@@ -26,6 +29,9 @@
     },
     {
       "path": "packages/lib/pod/tsconfig.cjs.json"
+    },
+    {
+      "path": "packages/lib/server-shared/tsconfig.cjs.json"
     },
     {
       "path": "packages/lib/util/tsconfig.cjs.json"
